requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0
yfinance==0.2.40
anthropic==0.34.0
openai==1.45.0
cryptography==43.0.0
pywebview==5.3.2
